My first project in python
In the above code we are going to take an input from user called user_string, and then we  use split function in  our string and then i declare a variable a whic store the acronyms of a phrase then we run a for loop in string and print the 0 index value every word after a split and turn it into an uppercase formate by usinf upper() function. and then print the achronyms of phrase.